adding execution component
functioning code/model instead of preprocessing
