+++ .gitattributes
@@ -1,0 +1,2 @@
+*.bin filter=lfs diff=lfs merge=lfs -text
+'*.BIN' filter=lfs diff=lfs merge=lfs -text
