+++ rel_component/TG-Relation-Model/README.md
@@ -1,0 +1,3 @@
+---
+license: afl-3.0
+---
