+++ rel_component/TG-Relation-Model/meta.json
@@ -1,0 +1,49 @@
+{
+  "lang":"en",
+  "name":"pipeline",
+  "version":"0.0.0",
+  "spacy_version":">=3.2.4,<3.3.0",
+  "description":"",
+  "author":"",
+  "email":"",
+  "url":"",
+  "license":"",
+  "spacy_git_version":"b50fe5ec6",
+  "vectors":{
+    "width":0,
+    "vectors":0,
+    "keys":0,
+    "name":null,
+    "mode":"default"
+  },
+  "labels":{
+    "transformer":[
+
+    ],
+    "relation_extractor":[
+      "CARD INTEREST RATE",
+      "LOCATED AT",
+      "ISSUED BY",
+      "RELATED",
+      "JURISDICTION IN"
+    ]
+  },
+  "pipeline":[
+    "transformer",
+    "relation_extractor"
+  ],
+  "components":[
+    "transformer",
+    "relation_extractor"
+  ],
+  "disabled":[
+
+  ],
+  "performance":{
+    "rel_micro_p":1.0,
+    "rel_micro_r":1.0,
+    "rel_micro_f":1.0,
+    "transformer_loss":0.000000016,
+    "relation_extractor_loss":0.0001716636
+  }
+}
