+++ rel_component/TG-Relation-Model/transformer/cfg
@@ -1,0 +1,3 @@
+{
+  "max_batch_items":4096
+}
