+++ rel_component/TG-Relation-Model/transformer/model/config.json
@@ -1,0 +1,65 @@
+{
+  "_name_or_path": "timhbach/Team-Gryffindor-bert-base-finetuned-NER-creditcardcontract",
+  "_num_labels": 9,
+  "architectures": [
+    "BertForTokenClassification"
+  ],
+  "attention_probs_dropout_prob": 0.1,
+  "classifier_dropout": null,
+  "hidden_act": "gelu",
+  "hidden_dropout_prob": 0.1,
+  "hidden_size": 768,
+  "id2label": {
+    "0": "O",
+    "1": "B-MISC",
+    "2": "I-MISC",
+    "3": "B-PER",
+    "4": "I-PER",
+    "5": "B-ORG",
+    "6": "I-ORG",
+    "7": "B-LOC",
+    "8": "I-LOC",
+    "9": "B-Issuer",
+    "10": "I-Issuer",
+    "11": "B-Card Name",
+    "12": "I-Card Name",
+    "13": "B-Jurisdiction",
+    "14": "I-Jurisdiction",
+    "15": "B-Interest Rate",
+    "16": "I-Interest Rate"
+  },
+  "initializer_range": 0.02,
+  "intermediate_size": 3072,
+  "label2id": {
+    "B-Card Name": "11",
+    "B-Interest Rate": "15",
+    "B-Issuer": "9",
+    "B-Jurisdiction": "13",
+    "B-LOC": "7",
+    "B-MISC": "1",
+    "B-ORG": "5",
+    "B-PER": "3",
+    "I-Card Name": "12",
+    "I-Interest Rate": "16",
+    "I-Issuer": "10",
+    "I-Jurisdiction": "14",
+    "I-LOC": "8",
+    "I-MISC": "2",
+    "I-ORG": "6",
+    "I-PER": "4",
+    "O": "0"
+  },
+  "layer_norm_eps": 1e-12,
+  "max_position_embeddings": 512,
+  "model_type": "bert",
+  "num_attention_heads": 12,
+  "num_hidden_layers": 12,
+  "output_past": true,
+  "pad_token_id": 0,
+  "position_embedding_type": "absolute",
+  "torch_dtype": "float32",
+  "transformers_version": "4.17.0",
+  "type_vocab_size": 2,
+  "use_cache": true,
+  "vocab_size": 28996
+}
