+++ rel_component/TG-Relation-Model/transformer/model/special_tokens_map.json
@@ -1,0 +1,1 @@
+{"unk_token": "[UNK]", "sep_token": "[SEP]", "pad_token": "[PAD]", "cls_token": "[CLS]", "mask_token": "[MASK]"}
