+++ rel_component/TG-Relation-Model/transformer/model/tokenizer_config.json
@@ -1,0 +1,1 @@
+{"do_lower_case": false, "unk_token": "[UNK]", "sep_token": "[SEP]", "pad_token": "[PAD]", "cls_token": "[CLS]", "mask_token": "[MASK]", "tokenize_chinese_chars": true, "strip_accents": null, "max_len": 512, "special_tokens_map_file": "C:\\Users\\anhhb/.cache\\huggingface\\transformers\\8ecdadef2bc275e74e0d4541ae8a5db151fba13174b86dfa88ef5765d30feb77.dd8bd9bfd3664b530ea4e645105f557769387b3da9f79bdb55ed556bdd80611d", "name_or_path": "timhbach/Team-Gryffindor-bert-base-finetuned-NER-creditcardcontract", "do_basic_tokenize": true, "never_split": null, "tokenizer_class": "BertTokenizer"}
