+++ rel_component/TG-Relation-Model/vocab/key2row
@@ -1,0 +1,1 @@
+€
