+++ rel_component/TG-Relation-Model/vocab/vectors.cfg
@@ -1,0 +1,3 @@
+{
+  "mode":"default"
+}
